Grammar police
Grammar police report: 
We solved the case of the missing M in "recommendations" spotted on "Why is this so important?" section, third dot.
Be aware, its said that followers are keeping the work of this evil maniac even after we captured him... more M's could be missing, maybe a "N" is added... !h! God! an  apostrophe could be missuse!!!!
Thanks and keep up the good work!