Updated english ver.
Deleted word "just" in the phrase "they are just he only way we ...", redundant expression when both words are written. Could also be "they are the only way", but as for the intention of the phrase thought "they are just the way we have..." fits best.
Changed the word "things" for "tasks", as for implying the execution of something, the word task fits better than things.
Added a comma after "To start" in the variable declaration paragraph.
Deleted "s" character in the phrase "...main differences of this variables...scope", due to it is addressing only one main difference (scope).
Added "\'s" to the word "variable" in the phrase "To retrieve a variable value", because we want to address to that variable's value, therefore is needed.