Update intro-to-prework.md
The human brain requires a lot of energy to make any decision, especially when lots of variables are taking part in the equation.  This course will be hard because we are re-designing the way you make these decisions. You will have to think similar to the way computers think. Do not worry though, a time will come when everything you do will become an acquired behaviour.

put a "u"