Update css-layouts.md
update changes to the sentence from the content under Top Menu with Left Sidebar.
from: This is almost the same, but instead of having the left-sidebar as the primary menu, you will use it for complementary content.  The main menu should be at the top box because is the first thing users see when they land on the website.

to: This is almost the same, but instead of having the left-sidebar as the primary menu, you will use it for complementary content.  The main menu should be at the top box because it is the first thing users see when they land on the website.