Spell check project
Original Snippet 1:
"Python was meant to be simple and easy. Here is the Python manifest: https://en.wikipedia.org/wiki/Zen_of_Python"
Corrected Snippet 1:
"Python was meant to be simple and easy. Here is the Python manifesto: https://en.wikipedia.org/wiki/Zen_of_Python"

Original Snippet 2:
"In Python, you can make any folder, a package by creating a '\__init\__.py' file inside of it."
Corrected Snippet 2:
"In Python, you can make any folder a package by creating an '__init__.py' file inside of it."

Original Snippet 3:
With JavaScript, you were tied and limited to the browser, you can’t access the client’s computer, and it is basically a rendering language. But Python is different
Corrected Snippet 3:
With the advent of Node.js, JavaScript now also encompasses server-side capabilities. However, with JavaScript you were historically tied and limited to the browser, unable to access the client’s computer, as it was primarily seen as a rendering language. But Python was different