Update what-is-learnpack.md Latam-pt30-Project
I changed the first part of the 'Note' content where you can read 'Note: If that does not work, Try...'
I replaced the 'does not work' to "doesn't work'. I believe it's easier to read it like this, specially if you are a Spanish native student trying to learn in the English Version.

Great Read btw, thanks for your hard work!!! Cheers.